working settings update. Settings are loaded from the DB and added to the AppConfig during startup. When updating settings, they are stored in AppConfig, and written do  the database.

+++ webapp/backend/pkg/models/setting_entry.go
@@ -13,7 +13,7 @@ type SettingEntry struct {
 	SettingKeyDescription string `json:"setting_key_description"`
 	SettingDataType       string `json:"setting_data_type"`
 
-	SettingValueNumeric int64  `json:"setting_value_numeric"`
+	SettingValueNumeric int    `json:"setting_value_numeric"`
 	SettingValueString  string `json:"setting_value_string"`
 }
